[NEW] Migrate IME to wayland.
This is to fix an issue in Firefox under Gnome (Wayland). When setting
IME to ibus, the position calculation is wrong, causing input method
position out of screen.

As a side effect, Chrome stops working in Gnome + Wayland until we set
IME module to ibus. There's no easy solution yet.